CONTRIB: rm Vagrantfile, instead use TravisCI config
The non-generic parts of the Vagrantfile are for building and
installation within a new Ubuntu 14.04 virual machine (VM).

Achieving these is prerequisite to testing, i.e., the same purposes
are served by part of the Travis CI configuration for TuLiP. As such,
the Travis CI configuration (or rather, code for CI testing of TuLiP)
should be re-used by other scripts that bootstrap VMs for general
purposes.

The remainder of the Vagrantfile after removing the custom install
script is not worth keeping because it is approximately the same as
the generic Vagrantfile obtained from `vagrant init`.

[ci skip]